Move helpers from CiviUnitTestCase to ContactTestTrait
* `createLoggedInUser(...)`
* `organizationCreate(...)`
* `individualCreate(...)`
* `householdCreate(...)`
* `sampleContact(...)`
* `_contactCreate(...)`
* `contactDelete(...)`
* `groupContactCreate(...)`
* `groupCreate(...)`
* `groupDelete(...)`
